Refactor nfce service
