+++ src/main/java/com/prati/projetomercado/controller/NfceController.java
@@ -2,13 +2,20 @@ package com.prati.projetomercado.controller;
 
 import com.prati.projetomercado.dto.request.NfceDataRequest;
 import com.prati.projetomercado.dto.response.SuccessResponse;
-import com.prati.projetomercado.entity.Purchase;
-import com.prati.projetomercado.service.NfceService;
+import com.prati.projetomercado.service.nfce.NfceService;
 import lombok.Getter;
 import lombok.RequiredArgsConstructor;
 import lombok.Setter;
 import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.*;
+import org.springframework.web.bind.annotation.DeleteMapping;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.PathVariable;
+import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.PutMapping;
+import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.RequestHeader;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RestController;
 
 import java.util.List;
 
@@ -16,26 +23,14 @@ import java.util.List;
 @RequestMapping("/api/nfce")
 @RequiredArgsConstructor
 public class NfceController {
-
+    
     private final NfceService nfceService;
-
-    @Setter
-    @Getter
-    public static class UrlRequest {
-        private String url;
-    }
-
-    @Setter
-    @Getter
-    public static class AccessKeyRequest {
-        private String accessKey;
-    }
 
     @GetMapping("/all")
     public ResponseEntity<SuccessResponse<List<NfceDataRequest>>> getAll(@RequestHeader("Authorization") String authorization) {
         List<NfceDataRequest> data = nfceService.getAll(authorization);
 
-        if(data.isEmpty()) {
+        if (data.isEmpty()) {
             return ResponseEntity.ok(new SuccessResponse<>("Nenhuma nota fiscal encontrada."));
         }
 
@@ -50,29 +45,41 @@ public class NfceController {
 
     @PostMapping("/register-link")
     public ResponseEntity<SuccessResponse<NfceDataRequest>> registerLink(@RequestHeader("Authorization") String authorization,
-                                             @RequestBody UrlRequest request) {
+                                                                         @RequestBody UrlRequest request) {
         NfceDataRequest data = nfceService.registerLink(authorization, request.getUrl());
         return ResponseEntity.ok(new SuccessResponse<>("Nota fiscal pelo link cadastrada com sucesso.", data));
     }
 
     @PostMapping("/register-manual")
     public ResponseEntity<SuccessResponse<Void>> registerManual(@RequestHeader("Authorization") String authorization,
-                                                 @RequestBody NfceDataRequest manualData) {
+                                                                @RequestBody NfceDataRequest manualData) {
         nfceService.registerManual(authorization, manualData);
         return ResponseEntity.ok(new SuccessResponse<>("Nota fiscal manual cadastrada com sucesso."));
     }
 
     @PutMapping("/edit")
     public ResponseEntity<SuccessResponse<Void>> edit(@RequestHeader("Authorization") String authorization,
-                                                          @RequestBody NfceDataRequest updatedNfce) {
+                                                      @RequestBody NfceDataRequest updatedNfce) {
         nfceService.edit(authorization, updatedNfce);
         return ResponseEntity.ok(new SuccessResponse<>("Nota fiscal editada com sucesso."));
     }
 
     @DeleteMapping("/delete")
     public ResponseEntity<SuccessResponse<Void>> delete(@RequestHeader("Authorization") String authorization,
-                                         @RequestBody AccessKeyRequest request)  {
+                                                        @RequestBody AccessKeyRequest request) {
         nfceService.delete(authorization, request.accessKey);
         return ResponseEntity.ok(new SuccessResponse<>("Nota fiscal deletada com sucesso."));
+    }
+
+    @Setter
+    @Getter
+    public static class UrlRequest {
+        private String url;
+    }
+
+    @Setter
+    @Getter
+    public static class AccessKeyRequest {
+        private String accessKey;
     }
 }
