+++ src/main/java/com/prati/projetomercado/repository/PurchaseRepository.java
@@ -1,5 +1,6 @@
 package com.prati.projetomercado.repository;
 
+import com.prati.projetomercado.entity.AuthUser;
 import com.prati.projetomercado.entity.Purchase;
 import org.springframework.data.jpa.repository.JpaRepository;
 
@@ -8,5 +9,6 @@ import java.util.Optional;
 
 public interface PurchaseRepository extends JpaRepository<Purchase, Long> {
     Optional<Purchase> findByAccessKey(String accessKey);
-    List<Purchase> findAllByUserId(Long userId);
+
+    List<Purchase> findAllByUser(AuthUser user);
 }
